VxWorks - Allow proper compilation
While on Linux system stat_vfs() and fstat_vfs() are used to gain
information about a mounted volume, VxWorks uses dirLib routines
statfs() and fstatfs(), that are POSIX-compliant routines to gain
information about a file system. This change includes the proper header
for VxWorks to to allow the proper access for those functions.