Changed the way path exports are declared. Some code cleanup.
Instead of exporting the whole path class, export only the functions that
are actually implemented in the shared library and leave the rest inline.
Also, move the definition of the static constants of the path class to
the header to avoid problems with exporting them from the shared library,
as observed with various MSVC versions and also gcc 7.3 from MinGW-w64.

Also, reduced code duplication and added a more efficient operator/
implementation when rvalue references are available.