Set S_IWUSR perm on the file created by copy_file and copy source perms on completion.
This ensures that the files overwritten or created by copy_file have the same
permission bits as the source file, as required by C++20. Also, for the duration
of the copy operation we ensure the target file has writing permission set.
This is important e.g. for NFS, which checks the permission on the server,
so a writable file descriptor on the client is not enough for the write
operation to succeed.

Notably, this doesn't save the case of overwriting the file with no write
permission set. In this case the operation will fail with EPERM.

Also, use fsync/fdatasync to guarantee that the target file is written
completely without errors before closing the file descriptor.