Marked filesystem_error with VISIBLE and exported members with DECL macros.
This makes a difference in case if Boost.Filesystem is linked statically into
user's application. In this case filesystem_error used to be not marked as
publicly visible, which could make it impossible to catch the exception if
it crossed shared library boundary. By marking the class as publicly visible
we ensure that RTTI is always visible, even in static builds.

The exported members are also marked with dllexport/dllimport attributes
for compatibility with Windows.