Explicitly qualify create_directory call to avoid ambiguity in path ctor.
As the compiler considers filesystem::create_directory overload that takes
path as the second argument, if NULL is defined as nullptr the compiler
attempts to construct path from nullptr, which is ambiguous and causes
compilation error. Remove that overload from the overload set by explicitly
qualifying the call.

Reported in https://github.com/boostorg/filesystem/pull/260.