Relax access rights for GetFileTime on Windows.
GetFileTime is documented to require GENERIC_READ access right, but this causes
problems if the file is opened by another process without FILE_SHARE_READ.
In practice, FILE_READ_ATTRIBUTES works, and FILE_READ_EA is also added for
good measure, in case if it matters for SMBv1.

If this doesn't work in some case, we might switch to
GetFileInformationByHandle(Ex) in the future.

Fixes https://github.com/boostorg/filesystem/issues/290.